refactor(proxy): extract tool injection tracker
